Making glue only 23b long

+++ cryptominisat4/searcher.cpp
@@ -152,7 +152,7 @@ void Searcher::add_lit_to_learnt(
                 if (varData[var].reason.getType() == clause_t) {
                     Clause* cl = cl_alloc.ptr(varData[var].reason.get_offset());
                     if (cl->red()) {
-                        lastDecisionLevel.push_back(std::make_pair(lit, cl->stats.glue));
+                        lastDecisionLevel.push_back(std::make_pair(lit, (uint32_t)cl->stats.glue));
                     }
                 } else if (varData[var].reason.getType() == binary_t
                     && varData[var].reason.isRedStep()
